Success 1388 Last Digit of A ^ B
Try 2 - 990746 Accepted

http://coj.uci.cu/24h/problem.xhtml?pid=1388
http://coj.uci.cu/24h/submission.xhtml?id=990746

Note: just fixed a little mistake.

+++ src/coj/java/P1388.java
@@ -26,10 +26,10 @@ public class P1388 {
                 System.out.println(a);
             } else {
                 
-                byte lastDigit = (byte)(a % (byte)10);            
+                byte lastDigit = (byte)(a % 10);            
                 byte[] pattern = getLastDigitsPattern(lastDigit);
 
-                byte remainder = (byte)(exp - 1 % 4);        
+                byte remainder = (byte)((exp - 1) % 4);        
                 byte index = (byte)((remainder + 3) % 4);
                 
                 System.out.println(pattern[index]);
